+++ src/oas/nodes/propObj.ts
@@ -82,7 +82,7 @@ export class PropObj extends Prop {
       writer.append(' '.repeat(context.indent + context.stack.length)).append('}');
     }
     // writer.append('\n');
-    if (context.generateOptions.debugParentInSelection) {
+    if (context.generateOptions.showParentInSelections) {
       writer.append(' # ').append(Naming.getRefName(this.parent!.name));
     }